Rust output may now be regression tested.
The output is now able to go through all syntax regression files.
It still spits out lots of warnings, but those can be worked on.

The regression suite works in the rust directory after one has run
`cargo fetch`.  It puts the output into the library and builds in
library mode.