Proper exception management for insert/update.
This makes it explicitly clear, both in code and in the documentation,
which errors can occur in using ort-nodejs(1).  Prior to this, it was
a bit too vague.

Updating and inserting now filter the database error for specific
constraint failures.

Moreover, exception handling is documented in ort-nodejs(1) to make
it easier for callers to assess exceptions.

This makes the behaviour of ort-c-source(1) and ort-nodejs(1) much
more similar with respect to error conditions for update and insert.