Allow nodejs/rust to override bcrypt cost.
When the "main" instance is created in either ort-json or ort-rust,
allow the number of bcrypt rounds ("cost") to be overridden.  Without
this, rust will default to 12 and nodejs to 10, making them both very
very slow for regression purposes.

Convert all tests to use 4 rounds.

Also add a test to both rust and nodejs that makes sure setting a
password hashed with OpenBSD's encrypt(1) is recognised by both tools.