Use "ort-paths.h" instead of "paths.h".
When compiling on systems not having readpassphrase() with the
regression utility using oconfigure (upcoming checkins...), the
compats.c file will try to include <paths.h>, which is a standard
header file in /usr/include.  Having "paths.h" in the local search
path causes the wrong header file to be included.